third_party/WebKit/LayoutTests/fast/hidpi/image-srcset-invalid-descriptor.html

<html>
<head>
<script>
    window.targetScaleFactor = 1;
</script>
<script src="resources/srcset-helper.js"></script>
<script src="../../resources/js-test.js"></script>
<script>
    addEventListener("load", function() {
        shouldBeTrue('document.getElementById("foo").clientWidth==200');
        shouldBeTrue('document.getElementById("foo2").clientWidth==200');
    }, false);
</script>
</head>

<body id="body">
    <div>This test passes if the image below is a 200px green square when the deviceScaleFactor is 1.</div>
    <div>It tests that the 1x resource is ignored since it contains invalid descriptors.</div>
    <div>The 2x resource has an intrinsic width of 200px, since it's a 400px wide image at 2x DPR.</div>
    <img id="foo" src="" srcset="resources/blue-100-px-square.png 43q 1x dfs 3dd, resources/green-400-px-square.png 2x">
    <img id="foo2" src="" srcset="resources/blue-100-px-square.png 300q, resources/green-400-px-square.png 2x">
</body>
</html>
