third_party/WebKit/LayoutTests/fast/hidpi/image-srcset-invalid-descriptor-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS document.getElementById("foo").clientWidth==200 is true
PASS document.getElementById("foo2").clientWidth==200 is true
This test passes if the image below is a 200px green square when the deviceScaleFactor is 1.
It tests that the 1x resource is ignored since it contains invalid descriptors.
The 2x resource has an intrinsic width of 200px, since it's a 400px wide image at 2x DPR.
  
